Upload modifications for final paper submission

+++ user_service/src/main/java/com/mercado_liebre/user_service/model/userAddress/UserAddressCreateDTO.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigInteger;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -15,6 +17,6 @@ public class UserAddressCreateDTO {
     private String province;
     private String city;
     private int postalCode;
-    private int contactPhone;
+    private BigInteger contactPhone;
     private UserDTO user;
 }
